[Android] Added Edit Profile Button and View
Profile button has an onclick listener that will redirect you to an
empty page, this is where the profile details page will be. I also fixed
the focus bug and keyboard bug on the main screen. Now when you launch
into the main activity the search bar is not focused and the keyboard is
not shown.