[Android] Added Edit Brewery Button
This button will only show if the userId logged in is a brewery manager.
If they are not a brewery manager then it will be hidden from view and
not clickable. Using a callback here to return from an async call to
determine if the current user is a brewery manager.