[Android] Changed JSON Converter, More Ratings
I had to change the JSON converter from Moshi to GSON as GSON better
supported custom adapters for dates which we needed to send the review
to the backend. I also added the star ratings, age picker, and fixed
some UI bugs with how the keyboard was hiding the 'comments' section
when editing text.